rust_bindings/python/test/requirements.txt
pytest==5.4.3
attrs>=17.4.0
iniconfig
more-itertools>=4.0.0
packaging
pluggy>=0.12,<1.0
py>=1.8.2
toml
importlib-metadata>=0.12;
wcwidth==0.2.5
PyYAML==5.3.1
